+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v2.0.0] - 2026-06-16
+
+### Added
+- File management (copy, paste, and delete) and the ability to name/rename your save slot.
+
+### Updated
+- previous and next arrows now gray out when there are no more save slot pages in that direction.
+
 ## [v1.0.0] - 2026-06-09
 
 ### Added